presentations/test-the-web-forward-overview/index.html

<!doctype html>  
<html lang="en">
	
	<head>
		<meta charset="utf-8">
		
		<title>Test the Web Forward</title>
        <meta name="apple-mobile-web-app-capable" content="yes" />
		<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
		
		<link href='http://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
		
		<link rel="stylesheet" href="css/reset.css">
		<link rel="stylesheet" href="css/main.css">
		<link rel="stylesheet" href="css/print.css" type="text/css" media="print">

		<link rel="stylesheet" href="lib/textmate.css">
		<link rel="stylesheet" href="css/adobe.css">
	</head>
	
	<body>
		
		<div class="reveal">

			<!-- Used to fade in a background when a specific slide state is reached -->
			<div class="state-background"></div>
			
			<!-- Any section element inside of this container is displayed as a slide -->
			<div class="slides">
				<section>
					<h2>Test the Web Forward</h2>
					<img src="assets/images/W3C_logo.jpg">
					<script>
						// Delicously hacky. Look away.
						if( navigator.userAgent.match( /(iPhone|iPad|iPod|Android)/i ) ) document.write( '<p style="color: rgba(0,0,0,0.3); text-shadow: none;">('+'Tap to navigate'+')</p>' );
					</script>
				</section>
				
				<section>
					<h2>Reftest Overview</h2>
					<blockquote>
					<ul>
					    <li class="fragment">Two files: Test file + Reference file</li>
					    <li class="fragment">Test file uses the feature you're test</li>
					    <li class="fragment">Reference file is an exact visual match without using the feature you're testing'</li>
						<li class="fragment">Self-describing (works both manually and with automation)</li>
						<li class="fragment"> Cross-browser & cross-platform</li>
					</ul>
					</blockquote>
				</section>
				
				<section>
    				<h2>Directory structure</h2>
    				<section>
    				    <br>
    				    
    				    <ul>
    				    <li>CSSWG Test Repository</li>
    				    <ul>
    				        <li>test</li>
    				        <ul>
    				            <li>contributors</li>
						        <ul>
						            <li>ttwf</li>
						            <ul>
						                <li>your_csswg_id</li>
						                <ul>
						                    <li>incoming</li>
    						                <li>submitted</li>
						                </ul>
						            </ul>
						        </ul>
						    </ul>
						</ul>
						</li>
					</section>
					<section>
					    <img src="assets/images/directoryStructure.png"/>
					</section>
    			</section>
    			<section>
    				<h2>Sample Test Case</h2>
    				<section>
    					<br><br>
    					<p style="font-size: 28pt">Basic test for the transform property described in the CSS3 Transforms spec using the translate() function</p>
    					<br>
    					<br>
    					<p>Spec Links:</p>
    					<p style="font-size: 18pt">
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#transform-property" target="top">http://dev.w3.org/csswg/css3-transforms/#transform-property</a>
    					    <br>
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
					    </p>
					</section>
					<section>
					    <img src="assets/images/specAssertion.png">
					</section>
	            </section>
				
				<section data-state="screenshotPage">
    			    <h3>Sample files</h3>
    			    <p>Test</p>
    				<img src="assets/images/001a.png"/>
    			    <p>Reference</p>
    			    <img src="assets/images/001b.png"/>
    			</section>
			
				<section data-state="screenshotPage">
					<h3>Examine the code</h3>
					<section>
					    <p>Test file</p>
					    <pre>
					    <code class="html">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
    &lt;head&gt;
    &lt;title&gt;CSS Transforms Test: transform property with translate function&lt;/title&gt;
    &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;
    &lt;link rel="match" href="reftest/ttwf-reftest-tutorial-ref.html"&gt;
    &lt;meta name="assert" content="This transform moves the element by 100 pixels in both the X and Y directions."&gt;
    &lt;<span class="styleTag">style type</span>="text/css"&gt;
            <span class="styleTag">.greenSquare</span> {
                <span class="cssProperty">position: absolute;
                top: 0;
                left: 0;
                width: 100<span class="px">px</span>;
                height: 100<span class="px">px</span>;
                background: green;</span>
                transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);
            }
            <span class="styleTag">.redSquare</span> {
                <span class="cssProperty">position: absolute;
                top: 101<span class="px">px</span>;
                left: 101<span class="px">px</span>;
                width: 98<span class="px">px</span>;
                height: 98<span class="px">px</span>;
                background: red;</span>
            }
        &lt;/<span class="styleTag">style</span>&gt;
    &lt;/head&gt;
    &lt;body&gt;
        &lt;p&gt;The test passes if there is a green square and no red.&lt;/p&gt;
        &lt;div class="redSquare"&gt;&lt;/div&gt;
        &lt;div class="greenSquare"&gt;&lt;/div&gt;
    &lt;/body&gt;
&lt;/html&gt;
                            </code>
					    </pre>
					</section>
					<section>
                        <p>Reference file</p>
                    	<pre>
                    	<code class="html">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
    &lt;head&gt;
    &lt;title&gt;CSS Transforms Reference file&lt;/title&gt;
    &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
    &lt;<span class="styleTag">style type</span>="text/css"&gt;
        <span class="styleTag">.greenSquare</span> {
            <span class="cssProperty">position: absolute;
            top: 100;
            left: 100;
            width: 100<span class="px">px</span>;
            height: 100<span class="px">px</span>;
            background: green;</span>
        }
    &lt;/<span class="styleTag">style</span>&gt;
    &lt;/head&gt;
        &lt;body&gt;
            &lt;p&gt;The test passes if there is a green square and no red.&lt;/p&gt;
            &lt;div class="greenSquare"&gt;&lt;/div&gt;
        &lt;/body&gt;
    &lt;/html&gt;
                                </code>
    					    </pre>
    					</section>
	        	</section>
				
				<section data-state="screenshotPage">
				    <h3>Required Metadata: Title</h3>
				    <br>
				    <p>Test file</p>
				    <pre>
					<code class="html">
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
    &lt;head&gt;
    &lt;title&gt;CSS Transforms Test: transform property with translate function&lt;/title&gt;
        . . .			
    </span>		        
				    </code>
				    </pre>
		            <p>Reference file</p>
		            <pre>
                    <code>
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
    &lt;head&gt;
    &lt;title&gt;CSS Transforms Reference file&lt;/title&gt;
        . . .			
</span>
                    </code>
                    </pre>
			    </section>
                
                <section data-state="screenshotPage">
    			    <h3>Required Metadata: Author</h3>
                    <br>
                    <p>Test file</p>
                    <pre>
                    <code class="html">
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    . . .			
    &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
    . . .			
</span>		        
				    </code>
				    </pre>
		            <p>Reference file</p>
                    <pre>
                    <code>
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    . . .
    &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
    . . .			
</span>
                    </code>
                    </pre>
                </section>
                
                <section data-state="expandPageMed">
    			    <h3>Required Metadata: Spec Links</h3>
    			    <p>Test file only</p>
                    <pre>
                    <code>
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    . . .
    &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;          . . .
</span>
                    </code>
                    </pre>
    			</section>

                <section data-state="expandPageMed">
    			    <h3>Required Metadata: Reference File Path</h3>
        		    <p>Test file only</p>
                    <pre>
                    <code>
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    . . .   
    &lt;link rel="match" href="reftest/ttwf-reftest-tutorial-ref.html"&gt;
    . . .
</span>
                        </code>
                        </pre>
    			</section>
    		    
    		    <section>
    		        <h3>Required Metadata: Assertion</h3>
                    <section data-state="expandPageWide">
            	        <p>Test file only</p>
                        <pre>
                        <code>
<span class="expand">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    . . .   
    &lt;meta name="assert" content="This transform moves the element by 100 pixels in both the X and Y directions."&gt;
    . . .
</span>
                        </code>
                        </pre>
    			    </section>
    			    <section>
			            <img src="assets/images/specAssertion.png">
			        </section>
    			</section>
    		
    			
    			<section>
    			    <h3>The Test</h3>
    			    <section>
    				    <p>Test element with the transform applied</p>
    				    <pre>
    				    <code class="html">
<span class="expand2">
                    &lt;<span class="styleTag">style type</span>="text/css"&gt;
                        <span class="styleTag">.greenSquare</span> {
                            <span class="cssProperty">position: absolute;
                            top: 0;
                            left: 0;
                            width: 100<span class="px">px</span>;
                            height: 100<span class="px">px</span>;
                            background: green;</span>
                            transform: translate(100<span class="px">px</span>,100<span class="px">px</span>);
                        }
                        . . .
                    &lt;/<span class="styleTag">style</span>&gt;    
</span>				        
    				    </code>
    				    </pre>
    			    </section>

    			    <section>
    				    <p>Reference element</p>
    				    <pre>
    				    <code class="html">
<span class="expand2">
                    &lt;<span class="styleTag">style type</span>="text/css"&gt;
                        <span class="styleTag">.greenSquare</span> {
                            <span class="cssProperty">position: absolute;
                            top: 100;
                            left: 100;
                            width: 100<span class="px">px</span>;
                            height: 100<span class="px">px</span>;
                            background: green;</span>
                        }
                    &lt;/<span class="styleTag">style</span>&gt;
</span>				        
    				    </code>
    				    </pre>
    			    </section>
                </section>
                
                <section>
    			    <h3>Visible failure</h3>
    			    <section>
    			        <p>Element is only visible when the test fails</p>
    			         <pre>
        				    <code class="html">
<span class="expand2">
                        &lt;<span class="styleTag">style type</span>="text/css"&gt;
                            . . .
                            <span class="styleTag">.redSquare</span> {
                                <span class="cssProperty">position: absolute;
                                top: 101<span class="px">px</span>;
                                left: 101<span class="px">px</span>;
                                width: 98<span class="px">px</span>;
                                height: 98<span class="px">px</span>;
                                background: red;</span>
                            }
                        &lt;/<span class="styleTag">style</span>&gt;    
</span>				        
        				    </code>
        				    </pre>
    			    </section>
    			    <section>
    				    <p>Failing test example</p>
    				    <img src="assets/images/failingRefTest.png"/>
    			    </section>
    			</section>
    			
    			<section>
    			    <h3>Self Description</h3>
    			    <p style="font-size: 24pt">Simple statement of how the page renders when the test passes</p>
    				<br>
    				<p style="font-size: 24pt"> Test file & Reference file </p>
    				 <pre>
    				 <code class="html">
<span class="expand2">
    &lt;body&gt;
        &lt;p&gt;The test passes if there is a green square and no red.&lt;/p&gt;
        . . .
    &lt;/body&gt;
</span>    				        
    			    </code>
    			    </pre>
                </section>
                
                <section data-state="screenshotPage">
        		    <h3>Add the elements</h3>
        		    <p>Test file</p>
        		     <pre>
    				 <code class="html">
<span class="expand2">
    &lt;body&gt;
        . . .
        &lt;div class="redSquare"&gt;&lt;/div&gt;
        &lt;div class="greenSquare"&gt;&lt;/div&gt;
    &lt;/body&gt;
</span>    				        
    			    </code>
    			    </pre>
    			    <p>Reference file</p>
                     <pre>
    				 <code class="html">
<span class="expand2">
    &lt;body&gt;
        . . .
        &lt;div class="greenSquare"&gt;&lt;/div&gt;
    &lt;/body&gt;
</span>    				        
    			    </code>
    			    </pre>
                </section>
    			
				<!-- Begin JS TEST -->
				<section data-state="largePage">
					<h2>JavaScript Test Overview</h2>
					<ul  style="line-height:50px">
					    <li>JavaScript tests programmatically verify functionality</li>
					    <li>JavaScript tests have some advantages over Ref Tests
					        <ul>
					            <li>Robust</li>
					            <li>Flexible</li>
					            <li>Performant</li>
					            <li>Powerful</li>
					        </ul>
					    </li>
					    <li>Ref Tests have some advantages over JavaScript tests
					        <ul>
					            <li>Tests verify complete rendering</li>
					            <li>Easy to understand</li>
					            <li>Avoids introducing additional technologies (JS)</li>
					        </ul>
					    </li>
					    <li>JavaScript tests can be used in conjunction with Ref Tests
					        <ul>
					            <li>Test can have both a reference file and a JavaScript section</li>
					            <li>Ref Test portion supports suites without script engines</li>
					        </ul>
					    </li>
					    <li>JavaScript tests best used in the following scenarios:
					        <ul>
					            <li>Testing a JS API or behavior (ex: CSSOM spec)</li>
					            <li>Adding automation to Ref Tests</li>
					        </ul>
					    </li>
					    <li>W3C provides a framework (testharness) to simplify and standardize JS test creation</li>
					</ul>
				</section>
				
				<section data-state="largePage">
				    <section>
				        <h3>testharness</h3>
				        <p>testharness.js and testharnessreport.js</p>
				        <pre style="font-size:24px">
				            <code class="html">
&lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;

&lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;
                            </code>
                        </pre>
				        
    				    <ul>
    				        <li>JavaScript framework that provides an API for making common test assertions</li>
    				        <li>Should be used by all JavaScript tests</li>
    				        <li>Path must be to a resources directory at the root level (/resources/...) prior to pushing to the test repository</li>
    				        <li>An element must exist in test file with id="log" to be used as a container for test results</li>
    				    </ul>
				    </section>
				    
			        <section>
    			        <h2>The test function</h2>
        			    <br>
        			    <pre style="font-size:24px">
        			        <code class="html">
                test(test_function, name, properties)
        			        </code>
        			    </pre>
        			    <ul>
        			        <li>Used for synchronous tests
        			            <ul>
        			                <li>async_test() is also provided by testharness for asychronous testing</li>
        			            </ul>
        			        </li>
        			        <li>Documented and implemented in testharness.js</li>
        			        <li>Used to define an individual test case in a file
        			            <ul>
        			                <li>Common for a file to have multiple calls to test()</li>
        			            </ul>
        			        </li>   
        			        <li>testharness will display PASS/FAIL for each call to test()</li>
        			         
        			    </ul>
                    </section>
                    
                    <section>
                        <h2>test() Argument 1: test_function</h2>
        			    <pre style="font-size:24px">
        			        <code class="html">
                test(<span style="color:red">test_function</span>, name, properties)
        			        </code>
        			    </pre>                        
                        <ul>
                            <li>The test_function argument must be a function object, not a function call</li>
                            <li>The test_function will typically be one of the assert methods supplied by testharness
                                <ul>
                                    <li>Can also be user defined if no assert method provides the needed functionality</li>
                                </ul>
                            </li>
                            <li>The evaluation of the assertion will determine PASS/FAIL for the test</li>
                            <br>
                            <li>Usage example:
                                <pre style="font-size:18px">
                                    <code>
                    test(function() {assert_true(true)}, name, properties)
                                    </code>
                                </pre>  
                            </li>                            
    			        </ul>                   
                        
                    </section>
                                 
                   <section>
                        <h2>test() Argument 2: Name</h2>
                        <pre style="font-size:24px">
                            <code class="html">
                test(test_function,<span style="color:red"> name</span>, properties)
                            </code>
                        </pre>                        
                        <ul>
                            <li>Name is a string that identifies the test</li> 
                            <li>Should be a descriptive test name that conveys what is being tested</li>
        			        <li>Name will be displayed in the test results table</li>
        			        <br>
                            <li>Usage example:
                                <pre style="font-size:18px">
                                    <code class="html">
            test(function() {assert_true(true)}, "Assert_true with true", properties)
                                    </code>
                                </pre>  
                            </li>
                        </ul>       			                                   
                    </section>

                    <section>
                        <h2>test() Argument 3: Properties</h2>
                        <pre style="font-size:24px">
                            <code class="html">
                test(test_function, name,<span style="color:red"> properties</span>)
                            </code>
                        </pre>                        
                        <ul>
                            <li>Properties is an object that overrides default test properties</li>
                            <li>timeout is the only currently recognized property
    			                <ul>
    			                    <li>timeout defines how long testharness will wait while executing a test case</li>
    			                    <li>This is specific to synchronous test, async_tests also have a timeout used slightly differently</li>
    			                </ul>
                            </li>
                            <br>
                            <li>Usage example:
                                <pre style="font-size:18px">
                                    <code class="html">
  test(function() {assert_true(true)}, "Assert_true with true", {timeout:1000})
                                    </code>
                                </pre>                                
                            </li>    			                    
    			        </ul>
                    </section>
                </section>
                  
        	    <section data-state="largePage">
        	        <section>
    			        <h2>Assert methods</h2>
        			    <ul>
        			        <li>testharness.js provides a number of assert methods to use for making test assertions
        			            <ul>
        			                <li>Full list is provided in the comment section of the testharness.js file</li>
        			                <li>Whenever possible, these methods should be used</li>
        			            </ul>
        			        </li>
        			        <li>All assert methods contain an optional "description" argument that is only output if the assertion fails
        			            <ul>
        			                <li>The description will be added to the default output from testharness</li>
        			                <li>The description should be used to provide additional debug information in the case of a failure</li>
        			            </ul>
        			        </li>
        			        <li>All asserts must be located in a test()
        			            <ul>
        			                <li>asserts outside won't be detected correctly by the harness and may cause a file to stop testing</li>
        			            </ul>
        			        </li>
        			    </ul>
    			    </section>
    			    
    			    <section>
    			        <h2>Assert Example</h2>
                        <pre style="font-size:24px">
                            <code class="html">
            assert_equals(actual, expected, description)
                            </code>
                        </pre>
                        <ul>
                            <li>actual: The actual value based on the functionality being tests</li>
                            <li>expected: The expected value, evaluated against actual value to determine if they are the same</li>
                            <li>description (optional): Output only if the assertion fails, used to provide additional debug info</li>
                            <br>
                            <li>Usage example:
                                <pre style="font-size:18px">
                                    <code class="html">
    assert_equals(getActualData("myElement"), 100, "Expected value for myElement is 100")
                                    </code>
                                </pre>                                
                            </li>                            
                        </ul>    			        
    			    </section>
				</section>

            	<section>
    				<h2>Sample Test Case</h2>
    				<section>
    					<br>
    					<p>Basic test for the transform property described in the CSS3 Transforms spec using the translate() function</p>
    					<br><br>
    					<p>Spec Links:</p>
    					<p style="font-size: 16pt">
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#transform-property" target="top">http://dev.w3.org/csswg/css3-transforms/#transform-property</a>
    					    <br>
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
					    </p>
					    <br>
					    <p style="color:red; font-style:italic; font-size: 16pt">Build upon Ref Test example to add JavaScript verification</p>
					</section>
					<section>
					    <img src="assets/images/specAssertion.png">
					</section>
	            </section>

				<section data-state="largePage">
				    <section>
				        <h3>Examine the code (head)</h3>
					    <pre style="font-size: 14px">
					        <code class="html">
&lt;!DOCTYPE html&gt;
&lt;html&gt;
&lt;head&gt;
    &lt;title&gt;CSS Transforms Test: translate function&lt;/title&gt;
    &lt;link rel="author" title="Your Name" href="mailto:youremail@address.com"&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property"&gt;
    &lt;link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions"&gt;
    &lt;link rel="match" href="reftest/ttwf-reftest-tutorial-ref.html"&gt;
    &lt;meta name="assert" content="This transform moves the element by 100 pixels in both the X and Y directions."&gt;
    &lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;
    &lt;script src="/resources/testharnessreport.js"&gt;&lt;/script&gt;    
    &lt;style type="text/css"&gt;
        * {
            margin: 0px;
            padding: 0px;
        }
        .results {
            position: absolute;
            top: 225px;
            width: 100%;
        }    
        .container {
            position: relative;
        }
        .greenSquare {
            position: absolute;
            width: 100px;
            height: 100px;
            background: green;
            transform: translate(100px,100px);
        }
        .redSquare {
            position: absolute;
            top: 101px;
            left: 101px;
            width: 98px;
            height: 98px;
            background: red;
        }
    &lt;/style&gt;
&lt;/head&gt;
                            </code>
                        </pre>
                    </section>
                
                    <section>
                        <h3>Examine the Code (Body)</h3>
                        <pre>
                            <code>
...
    &lt;body&gt;
        &lt;div class="results"&gt;
            &lt;p&gt;The test passes if there is a green square at 100,100 and no red is visible&lt;/p&gt;&lt;br&gt;
            &lt;div id="log"&gt;&lt;/div&gt;
        &lt;/div&gt;
        &lt;div class="container"&gt;
            &lt;div class="redSquare"&gt;&lt;/div&gt;
            &lt;div id="myDiv" class="greenSquare"&gt;&lt;/div&gt;
        &lt;/div&gt;
        &lt;script type="text/javascript"&gt;
            function checkElementPosition(elemID) {
                return document.getElementById(elemID).getBoundingClientRect().left + "," +
                       document.getElementById(elemID).getBoundingClientRect().top;            
            }    
            test(function() {assert_equals(checkElementPosition("myDiv"), "100,100", "myDiv should be at 100,100")}, 
                            "Transform_Translate_100px_100px");
        &lt;/script&gt;    
    &lt;/body&gt;
&lt;/html&gt;					    
					        </code>
					    </pre>
				    </section>
			    </section>				
			    
			    <section data-state="largePage">
			        <h3>Positioning</h3>
			        <pre>
			            <code class="html">
&lt;style type="text/css"&gt;
    * {
        margin: 0px;
        padding: 0px;
    }

    .results {
        position: absolute;
        top: 225px;
        width: 100%;
    }    
...
&lt;/style&gt;
                        </code>
                    </pre>
                    <br>
                    <ul>
                        <li>Not required for the test</li> 
                        <li>Simplifies the code by removing offsets caused by margins/padding</li>
                        <li>Results container positioned below test element to avoid interference</li>
                        <li>Test could also omit these styles and account for offsets in expected value</li>
                    </ul>
			    </section>
 			    
			    <section data-state="largePage">
				    <h3>User Defined Function</h3>
                    <pre>
                        <code class="html">
...
        function checkElementPosition(elemID) {
            return document.getElementById(elemID).getBoundingClientRect().left + "," +
                   document.getElementById(elemID).getBoundingClientRect().top;            
        }    
...
                        </code>
                    </pre>
                    <br>
                    <ul>
                        <li>User defined function to get the actual value of the element being tested</li>
                        <li>Uses the DOM function getBoundingClientRect() to get the x,y coordinates of an element based on its ID</li> 
                        <li>Returns a string that consists of "X,Y"</li>
                    </ul>
			    </section>

			    <section data-state="largePage">
				    <h3>Test Case</h3>
				    <p>test() with assert_equals</p>
                        <pre>
                            <code class="html">
&lt;body&gt;
... 
    test(function() {assert_equals(checkElementPosition("myDiv"), "100,100", "myDiv should be at 100,100")}, 
                    "Transform_Translate_100px_100px");
...    
&lt;/body&gt;
&lt;/html&gt;
                            </code>
                        </pre>
				    <ul>
				        <li>assert_equals(actual, expected, description)
				            <ul>
				                <li>actual = checkElementPosition("myDiv")</li>
				                <li>expected = "100,100"</li>
				                <li>description = "myDiv should be at 100,100"</li>
				            </ul>
				        </li>
				        <br>				        
				        <li>test(test_function, name, properties)
				            <ul>
				                <li>test_function = assert_equals(checkElementPosition("myDiv"), "100,100", "myDiv should be at 100,100")</li>
				                <li>name = "Transform_Translate_100px_100px"</li>
				                <li>properties = <span style="font-style:italic">not specified</span></li>
				            </ul>
				        </li>
				    </ul>
				        
			    </section>   
                
    			<section>
    			    <section data-state="largePage">
    				    <h3>Test Results</h3>
    				    <p>PASS</p>
    				    <img src="assets/images/jsTest_PASS.png"/>
    			    </section>
    			    
                    <section data-state="largePage">
    				    <h3>Test Results</h3>
    				    <p>FAIL</p>
    				    <img src="assets/images/jsTest_FAIL.png"/>
    			    </section>    			    
                </section>                                                    
    			
    			<section>
    			    <h1>Q & A</h1>
    			</section>
		
			</div>

			<!-- The navigational controls UI -->
			<aside class="controls">
				<a class="left" href="#">&#x25C4;</a>
				<a class="right" href="#">&#x25BA;</a>
				<a class="up" href="#">&#x25B2;</a>
				<a class="down" href="#">&#x25BC;</a>
			</aside>

			<!-- Displays presentation progress, max value changes via JS to reflect # of slides -->
			<div class="progress"><span></span></div>
			
		</div>

		<!-- Optional libraries for code syntax highlighting and classList support in IE9 -->
		<script src="lib/highlight.js"></script>
		<script src="lib/classList.js"></script>
		
		<script src="js/reveal.js"></script>
		
		<script>
			// Parse the query string into a key/value object
			var query = {};
			location.search.replace( /[A-Z0-9]+?=(\w*)/gi, function(a) {
				query[ a.split( '=' ).shift() ] = a.split( '=' ).pop();
			} );

			// Fires when a slide with data-state=customevent is activated
			Reveal.addEventListener( 'customevent', function() {
				alert( '"customevent" has fired' );
			} );

			// Fires each time a new slide is activated
			Reveal.addEventListener( 'slidechanged', function( event ) {
				// event.indexh & event.indexv
			} );

			Reveal.initialize({
				// Display controls in the bottom right corner
				controls: true,

				// Display a presentation progress bar
				progress: true,

				// If true; each slide will be pushed to the browser history
				history: true,

				// Loops the presentation, defaults to false
				loop: false,

				// Flags if mouse wheel navigation should be enabled
				mouseWheel: true,

				// Apply a 3D roll to links on hover
				rollingLinks: true,

				// UI style
				theme: query.theme || 'default', // default/neon

				// Transition style
				transition: query.transition || 'default' // default/cube/page/concave/linear(2d)
			});

			hljs.initHighlightingOnLoad();
		</script>

	</body>
</html>
